changed routes,done with some formatting,added prompt on delete all on both pages kra and role

+++ src/pages/Kra/kra.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { ToastContainer, toast } from 'react-toastify';
-import { environment } from './Environment'
+import { environment } from '../Environment';
 import bootbox from 'bootbox';
 const $ = require('jquery');
 $.DataTable = require('datatables.net-bs4');
@@ -57,32 +57,51 @@ class kraListPage extends Component {
             }
         });
     }
+
     DeleteAllKra() {
         $("#kraDataList input:checkbox:checked").each((e, item) => {
             this.state.selectedIds.push(item.value);
         });
         var res = '';
 
-        if (this.state.selectedIds.length > 0) {
-            this.state.selectedIds.map(item => {
-                res = this.DeleteKraApi(item);
-            });
-            res.done(response => {
-                toast.success("KRA Deleted Successfully !", {
-                    position: toast.POSITION.TOP_RIGHT
+      
+            if (this.state.selectedIds.length > 0) {
+            
+                bootbox.confirm({
+                    message: "Delete this record ?",
+                    buttons: {
+                        confirm: {
+                            label: 'Yes',
+                            className: 'btn-success'
+                        },
+                        cancel: {
+                            label: 'No',
+                            className: 'btn-danger'
+                        }
+                    },
+                    callback: (result) => {
+                        if (result === true) {
+                            this.state.selectedIds.map(item => {
+                                res = this.DeleteKraApi(item);
+                            });
+                            res.done(response => {
+                                toast.success("Kra Deleted Successfully !", {
+                                    position: toast.POSITION.TOP_RIGHT
+                                });
+                                this.$el.DataTable().ajax.reload();
+                            });
+                            res.fail(error => {
+                
+                            });
+                        }
+                    }
                 });
-                this.$el.DataTable().ajax.reload();
-            });
-            res.fail(error => {
-
-            });
-
-        } else {
-            toast.info("Select atleast one record", {
-                position: toast.POSITION.TOP_RIGHT
-            });
-        }
-    }
+                }
+                else {
+                    toast.info("please select atleast one record!");
+                }
+            }
+              
 
     SingleDeleteConfirm(id) {
         if (id !== undefined) {
@@ -112,30 +131,7 @@ class kraListPage extends Component {
     multiKraDeleteConfirm(id) {
 
 
-        //    if (id !== undefined) {
-        bootbox.confirm({
-            message: "Delete this record ?",
-            buttons: {
-                confirm: {
-                    label: 'Yes',
-                    className: 'btn-success'
-                },
-                cancel: {
-                    label: 'No',
-                    className: 'btn-danger'
-                }
-            },
-            callback: (result) => {
-                if (result === true) {
-                    this.DeleteAllKra(id);
-                }
-                else {
-                    toast.info("Select atleast one record", {
-                        position: toast.POSITION.TOP_RIGHT
-                    });
-                }
-            }
-        });
+        
     }
 
     componentDidMount() {
@@ -144,6 +140,7 @@ class kraListPage extends Component {
         this.$el.DataTable({
             "autoWidth": false,
             aaSorting: [[1, 'asc']],
+            aaSorting: [[2, 'asc']],
             ajax: {
                 url: endpointGET,
                 // url: "http://180.211.103.189:3000/api/kra_master/",
@@ -167,18 +164,24 @@ class kraListPage extends Component {
 
                 },
                 {
+                    data: null,
+                    targets: 1,
+                    "orderable": false,
+
+                },
+                {
                     data: "kraName",
-                    targets: 1
+                    targets: 2
                 },
                 {
                     data: "description",
-                    targets: 2
+                    targets: 3
                 },
 
                 {
                     data: "action",
                     "orderable": false,
-                    targets: 3,
+                    targets: 4,
                     render: function (data, type, row) {
                         return (
                             '<a href="/Editkra/id=' + row.kraId + '"class="btn mr-2 btn-edit btn-info btn-sm">' + '<i class="fa fa-pencil" aria-hidden="true"></i>' +
@@ -191,6 +194,10 @@ class kraListPage extends Component {
 
                 }
             ],
+            "fnRowCallback": function (nRow, aData, iDisplayIndex) {
+                $("td:eq(1)", nRow).html(iDisplayIndex + 1);
+                return nRow;
+            },
             initComplete: (settings, json) => {
                 //debugger;
                 // $(".btnDelete").on("click", e => {
@@ -213,19 +220,15 @@ class kraListPage extends Component {
         return (
             //#region list table kra    
             <div>
-                <h1>Kra</h1>
-                {
-                    this.props.location.state === "2222"
-                }
-                <div className="clearfix text-right mb-2">
-                    <Link to={{ pathname: '/add-kra', state: {} }} className="btn btn-primary btn-lg"><i className="fa fa-plus" aria-hidden="true"></i></Link>
+                <div className="clearfix d-flex align-items-center row page-title">
+                    <h2 className="col">Kra</h2>
+                    <div className="col text-right">
+                        <Link to={{ pathname: '/kra/add', state: {} }} className="btn btn-primary "><i className="fa fa-plus" aria-hidden="true"></i></Link>
+                    </div>
+                    <button className="btn btn-danger btn-multi-delete btnDeleteAll" onClick={() => {
+                            this.DeleteAllKra();
+                        }}><i className="fa fa-trash" aria-hidden="true"></i></button>
                 </div>
-                <button
-                    className="btn btn-danger mb-2 btnDeleteAll"
-                    onClick={() => {
-                        this.multiKraDeleteConfirm(this);
-                    }}><i className="fa fa-trash" aria-hidden="true"></i></button>
-
                 <table className="table table-striped table-bordered table-hover customDataTable"
                     id="kraDataList"
                     ref={el => (this.el = el)}>
@@ -237,6 +240,7 @@ class kraListPage extends Component {
                                     onClick={e => { this.checkall(e); }}
                                 />
                             </th>
+                            <th width="50">Sr.No</th>
                             <th>Name</th>
                             <th>Description</th>
                             <th width="90">Action</th>
